refactor(tests): les tests d'api sont tous ensembles, séparés par leur type (unit, integration)
