+++ packages/api/src/api/graphql/utilisateurs.test.integration.ts
@@ -1,10 +1,14 @@
-import { app } from './app'
-import { graphQLCall, queryImport, tokenCreate } from './_utils/index'
-import { userAdd } from '../src/knex/user-add'
+import { app } from '../../../tests/app'
+import {
+  graphQLCall,
+  queryImport,
+  tokenCreate
+} from '../../../tests/_utils/index'
+import { userAdd } from '../../knex/user-add'
 import request from 'supertest'
 import jwt from 'jsonwebtoken'
-import { dbManager } from './db-manager'
-import { IUtilisateur } from '../src/types'
+import { dbManager } from '../../../tests/db-manager'
+import { IUtilisateur } from '../../types'
 import { Administrations } from 'camino-common/src/administrations'
 
 console.info = jest.fn()
