+++ packages/api/tsconfig.json
@@ -29,7 +29,7 @@
       "tus-node-server": ["./src/@types/tus-node-server"]
     }
   },
-  "exclude": ["node_modules", "**/*.test.ts", "**/__mocks__/*"],
+  "exclude": ["node_modules", "**/__mocks__/*"],
   "include": ["src/**/*", "sources/**/*"],
   "typedocOptions": {
     "out": "docs",
